// src/store/modules/user/mutation-types.ts
/*
 * @version:  ;
 * @description:  ;
 * @autor: Full
 * @date: Do not edit
 */
export enum userMutationTypes {
    SET_USERINFO = 'SET_USERINFO',
    SET_AMOUNT = 'SET_AMOUNT',
    SET_COMBO = 'SET_COMBO',
    SET_PAYLIST = 'SET_PAYLIST',
    SET_PAYAMOUNTLIST = 'SET_PAYAMOUNTLIST',
    SET_RECORDRECHARGE = 'SET_RECORDRECHARGE',
    SET_NEWRECORDRECHARGE = 'SET_NEWRECORDRECHARGE',
    SET_WALLETAD = 'SET_WALLETAD',
    SET_SHAREQRDATA = 'SET_SHAREQRDATA',
    SET_SHAREQRDATARECORED = 'SET_SHAREQRDATARECORED',
}